API admin route and cleanup implemented.
